Add files via upload
Light Blow Arduino with Ino Create:

To program an Arduino, you can use the Arduino IDE or Ino Create, an online platform for Arduino development.
You write the code (usually in C/C++) to control the Arduino's behavior, such as blinking an LED or reading sensor data.
After writing the code, you compile it to create a build, which checks for errors and prepares the program for the Arduino.
Load Hex File:

After successfully compiling the code, the Arduino IDE or Ino Create generates a HEX file. This file contains the machine code that the Arduino can understand.
You upload (load) this HEX file to the Arduino board using a USB cable or another appropriate method. This transfers the program to the Arduino's microcontroller.
Run the Simulator:

If you want to test your code before uploading it to the physical Arduino board, you can use a simulator.
The simulator is a software tool that emulates the behavior of an Arduino, allowing you to run and test your code virtually on your computer.
This virtual simulation can help you identify any potential issues and debug the code before deploying it to the actual hardware.